Debugged successfully the code organization at this point.
