+++ Pokemon/Player.cpp
@@ -1,6 +1,5 @@
 #include "Player.h"
 #include "PokemonChoice.h"
-#include "PokemonType.h"
 #include "Utility.h"
 #include <iostream>
