+++ Pokemon/Player.h
@@ -1,4 +1,5 @@
 #include <string>
+#include "Pokemon.h"
 
 using namespace std;
