+++ Pokemon/Utility.h
@@ -1,10 +1,8 @@
-//#pragma once
-//Utility.h
+#pragma once
 //Global Functions used in project go here.
 /*
 "The static keyword makes a function belong to the class itself, not to any object.
 You can call it using the class name: Utility::clearConsole(); , without needing to create an instance."
-- Outscal.com
 
 */
 class Utility {
